+++ src/common/utils.f90
@@ -49,7 +49,7 @@ module utils
        filename_suffix_pos, filename_tslash_pos, filename_split, &
        linear_index, split_string, NEKO_FNAME_LEN, index_is_on_facet, &
        concat_string_array, extract_fld_file_index, neko_type_error, &
-       neko_type_registration_error
+       neko_type_registration_error, NEKO_VARNAME_LEN
 
 
 contains