+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Develop
 
+- Added the variable NEKO_VARNAME_LEN in `common/utils.f90` to set a fixed
+  size for `name` attributes in e.g. `field_t` and `vector_t`.
 - Added the `data_streamer` simulation component, allowing data streaming 
   with ADIOS2.
 - Fixed a bug (mu_msk) in `device_calc_force_array` in `force_torque.f90`.